% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datageneration}
\alias{datageneration}
\title{Generate Artificial Data by Simulation}
\usage{
datageneration(n)
}
\arguments{
\item{n}{The sample size}
}
\value{
A list containing the outcome vector, the treatment vector,
the instrumental vector, and the true instrumental exposure vector, and
the symmetric binary adjacency matrix.
}
\description{
The \code{datageneration()} function generates artificial ring-network data
by simulation.
The function is used in the package vignette.
}
\examples{
latenetwork::datageneration(n = 2000)

}
